feat(cocode-rs): add bash execution, context compaction, and error handling improvements
Shell Execution:
- Add comprehensive bash command execution with shell state management
- Implement shell snapshot/restore for environment preservation
- Support for command execution with proper PTY handling
- Add shell types and cleanup utilities

Context Compaction:
- Implement conversation compaction for long-running sessions
- Add compact config and compaction driver logic
- Support loop events for compaction progress tracking
- Enhance message history management

Error Handling Refactoring:
- Simplify error types across core/tools, core/api, core/context, core/prompt
- Streamline tool error handling with consistent patterns
- Fix error propagation in provider factory and retry logic

Subagent & Coordinator Improvements:
- Enhance subagent manager with improved lifecycle management
- Add coordinator manager enhancements for multi-agent execution
- Add tool context support for subagent spawning

+++ cocode-rs/app/cli/src/output.rs
@@ -45,7 +45,7 @@ pub fn print_session_start(session_id: &str, model: &str, provider: &str) {
 }
 
 /// Print turn completion summary.
-pub fn print_turn_summary(input_tokens: i32, output_tokens: i32) {
+pub fn print_turn_summary(input_tokens: i64, output_tokens: i64) {
     println!();
     println!("[Tokens: {input_tokens} in / {output_tokens} out]");
 }
